+++ src/ZenUI.Wpf/Controls/ZenAlert.cs
@@ -85,8 +85,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 获取或设置由提示级别提供的默认背景画刷。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public Brush ThemeBackground
+        internal Brush ThemeBackground
         {
             get { return (Brush)GetValue(ThemeBackgroundProperty); }
             set { SetValue(ThemeBackgroundProperty, value); }
@@ -95,14 +94,13 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="ThemeBackground"/> 依赖属性。
         /// </summary>
-        public static readonly DependencyProperty ThemeBackgroundProperty =
+        internal static readonly DependencyProperty ThemeBackgroundProperty =
             RegisterBrush(nameof(ThemeBackground));
 
         /// <summary>
         /// 获取或设置由提示级别提供的默认边框画刷。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public Brush ThemeBorderBrush
+        internal Brush ThemeBorderBrush
         {
             get { return (Brush)GetValue(ThemeBorderBrushProperty); }
             set { SetValue(ThemeBorderBrushProperty, value); }
@@ -111,14 +109,13 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="ThemeBorderBrush"/> 依赖属性。
         /// </summary>
-        public static readonly DependencyProperty ThemeBorderBrushProperty =
+        internal static readonly DependencyProperty ThemeBorderBrushProperty =
             RegisterBrush(nameof(ThemeBorderBrush));
 
         /// <summary>
         /// 获取或设置由提示级别提供的默认图标背景画刷。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public Brush ThemeIconBackground
+        internal Brush ThemeIconBackground
         {
             get { return (Brush)GetValue(ThemeIconBackgroundProperty); }
             set { SetValue(ThemeIconBackgroundProperty, value); }
@@ -127,7 +124,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="ThemeIconBackground"/> 依赖属性。
         /// </summary>
-        public static readonly DependencyProperty ThemeIconBackgroundProperty =
+        internal static readonly DependencyProperty ThemeIconBackgroundProperty =
             RegisterBrush(nameof(ThemeIconBackground));
 
         /// <summary>