+++ src/ZenUI.Wpf/Controls/ZenRadioButton.cs
@@ -94,8 +94,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 获取或设置由选中状态提供的默认背景画刷。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public Brush ThemeBackground
+        internal Brush ThemeBackground
         {
             get { return (Brush)GetValue(ThemeBackgroundProperty); }
             set { SetValue(ThemeBackgroundProperty, value); }
@@ -104,14 +103,13 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="ThemeBackground"/> 依赖属性。
         /// </summary>
-        public static readonly DependencyProperty ThemeBackgroundProperty =
+        internal static readonly DependencyProperty ThemeBackgroundProperty =
             RegisterBrush(nameof(ThemeBackground));
 
         /// <summary>
         /// 获取或设置由选中状态提供的默认边框画刷。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public Brush ThemeBorderBrush
+        internal Brush ThemeBorderBrush
         {
             get { return (Brush)GetValue(ThemeBorderBrushProperty); }
             set { SetValue(ThemeBorderBrushProperty, value); }
@@ -120,7 +118,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="ThemeBorderBrush"/> 依赖属性。
         /// </summary>
-        public static readonly DependencyProperty ThemeBorderBrushProperty =
+        internal static readonly DependencyProperty ThemeBorderBrushProperty =
             RegisterBrush(nameof(ThemeBorderBrush));
 
         private static DependencyProperty RegisterBrush(string name)