+++ src/ZenUI.Wpf/Controls/ZenSwitch.cs
@@ -1,5 +1,4 @@
 using System;
-using System.ComponentModel;
 using System.Windows;
 using System.Windows.Automation.Peers;
 using System.Windows.Controls.Primitives;
@@ -77,8 +76,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 获取或设置由开关状态提供的默认轨道背景画刷。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public Brush ThemeBackground
+        internal Brush ThemeBackground
         {
             get { return (Brush)GetValue(ThemeBackgroundProperty); }
             set { SetValue(ThemeBackgroundProperty, value); }
@@ -87,7 +85,7 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="ThemeBackground"/> 依赖属性。
         /// </summary>
-        public static readonly DependencyProperty ThemeBackgroundProperty =
+        internal static readonly DependencyProperty ThemeBackgroundProperty =
             DependencyProperty.Register(
                 nameof(ThemeBackground),
                 typeof(Brush),
@@ -106,14 +104,13 @@ namespace ZenUI.Wpf.Controls
         /// <summary>
         /// 标识 <see cref="CapsuleCornerRadius"/> 只读依赖属性。
         /// </summary>
-        public static readonly DependencyProperty CapsuleCornerRadiusProperty =
+        internal static readonly DependencyProperty CapsuleCornerRadiusProperty =
             CapsuleCornerRadiusPropertyKey.DependencyProperty;
 
         /// <summary>
         /// 获取与开关实际高度相匹配的胶囊圆角。
         /// </summary>
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public CornerRadius CapsuleCornerRadius
+        internal CornerRadius CapsuleCornerRadius
         {
             get { return (CornerRadius)GetValue(CapsuleCornerRadiusProperty); }
         }